try-airflow/pyproject.toml
[tool.poetry]
name = "try-airflow"
version = "0.1.0"
description = ""
authors = ["po3rin <abctail30@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
apache-airflow = "^1.10.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
